use count_if to reduce loop

+++ puzzles/day02/day02.h
@@ -6,7 +6,7 @@
 
 bool is_safe(const std::vector<int> &level);
 
-int part1(const std::string &filename);
-int part2(const std::string &filename);
+long part1(const std::string &filename);
+long part2(const std::string &filename);
 
 #endif // AOC24_DAY01_H
